[arm64][turbofan] Remove incorrect DCHECK in macro assembler arm64
Bug: chromium:1454436
Reviewed-on: https://chromium-review.googlesource.com/c/v8/v8/+/4614432
Commit-Queue: Nico Hartmann <nicohartmann@chromium.org>
Cr-Commit-Position: refs/heads/main@{#88275}

+++ src/compiler/backend/arm64/instruction-selector-arm64.cc
@@ -122,7 +122,7 @@ class Arm64OperandGeneratorT final : public OperandGeneratorT<Adapter> {
       case kLogical32Imm:
         // TODO(dcarney): some unencodable values can be handled by
         // switching instructions.
-        return Assembler::IsImmLogical(static_cast<uint64_t>(value), 32,
+        return Assembler::IsImmLogical(static_cast<uint32_t>(value), 32,
                                        &ignored, &ignored, &ignored);
       case kLogical64Imm:
         return Assembler::IsImmLogical(static_cast<uint64_t>(value), 64,
